try: makefile

+++ Makefile
@@ -1,23 +1,43 @@
 VERSION ?= 1.0.0
 
-BLOOM_DEV_DOCKER = @docker run --name bloom-test  --mount type=bind,source="$(shell pwd)",target=/source_code --env-file ./.env.test --network=bloom_net -p 8501:8501
+BLOOM_DEV_DOCKER = docker run --name bloom-dev  --mount type=bind,source="$(shell pwd)",target=/source_code --env-file /tmp/docker.dev --network=bloom_net -p 8501:8501
 BLOOM_PRODUCTION_DOCKER = @docker run --mount type=bind,source="$(shell pwd)",target=/source_code --env-file ./.env --log-driver json-file --log-opt max-size=10M --log-opt max-file=3 --entrypoint /entrypoint.sh
 
-UPDATE_SETTINGS = @python -c "from bloom.config import settings"
+EXPORT_ENV_DEV = @export $(shell cat /tmp/docker.dev | grep -v "#" | xargs -d '\r')
+EXPORT_ENV_TEST = @export $(shell cat /tmp/docker.test | grep -v "#" | xargs -d '\r')
+EXPORT_ENV_PROD = @export $(shell cat /tmp/docker.prod | grep -v "#" | xargs -d '\r')
 
-# Regenerate .env file according to .env.template, .env.local, APP_ENV value, .env.${APP_ENV},  .env.${APP_ENV}.local
-update_env:
-	$(UPDATE_SETTINGS)
+UPDATE_ENV_DEV = @$(shell pwd)/scripts/update_env.sh dev /tmp/env.dev
+UPDATE_ENV_TEST = @$(shell pwd)/scripts/update_env.sh test /tmp/env.test
+UPDATE_ENV_PROD = @$(shell pwd)/scripts/update_env.sh prod /tmp/env.prod
+
+define set_env_dev
+$(1): APP_ENV=dev
+$(1): PATH_ENV=/tmp/docker.dev
+$(1): /venv/bin/python3 -c "from bloom.config import settings"
+endef
+
+export_dev:
+	@export APP_ENV=dev
+	@export APP_ENV=dev
+export $(cat /tmp/docker.$$APP_ENV | grep -v "#" | xargs -d '\r')
+export $(cat  /tmp/docker.$$APP_ENV | grep -v "#" | xargs -d '\r')
+
 	
 build:
 	@docker build -t d4g/bloom:${VERSION} --platform linux/amd64  -f docker-env/Dockerfile .
 	@docker tag d4g/bloom:${VERSION} d4g/bloom:latest
 
+test-config:
+	$(UPDATE_ENV_DEV)
+	$(EXPORT_ENV_DEV) && env 
+
 launch-dev-db:
-	@docker compose -f docker-env/docker-compose-db.yaml up -d
+	$(UPDATE_ENV_DEV)
+	$(EXPORT_ENV_DEV) && docker compose -f docker-env/docker-compose-db.yaml up -d
 	@sleep 20
-	$(BLOOM_DEV_DOCKER) --rm d4g/bloom:${VERSION} alembic upgrade head
-	$(BLOOM_DEV_DOCKER) --rm d4g/bloom:${VERSION} /venv/bin/python3 alembic/init_script/load_vessels_data.py
+	$(EXPORT_ENV_DEV) && $(BLOOM_DEV_DOCKER) --rm d4g/bloom:${VERSION} alembic upgrade head
+	$(EXPORT_ENV_DEV) && $(BLOOM_DEV_DOCKER) --rm d4g/bloom:${VERSION} /venv/bin/python3 alembic/init_script/load_vessels_data.py
 
 load-amp-data:
 	$(BLOOM_DEV_DOCKER) --rm d4g/bloom:${VERSION} /venv/bin/python3 alembic/init_script/load_amp_data.py
